% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-geoheat.R
\name{theme_geoheat}
\alias{theme_geoheat}
\title{Base geoheat theme}
\source{
Builds off of \code{theme_statebins()}, which builds off of \code{theme_bw()}
}
\usage{
theme_geoheat(legend_position = "right", base_size = 12, base_family = "")
}
\arguments{
\item{legend_position}{fills in \code{legend.position}}

\item{base_family, base_size}{same as \code{theme_bw()}}
}
\description{
Base geoheat theme
}

